ds binary search added

+++ Data Structure/Week-01/Module-03/Y_Range_sum_query.cpp
@@ -14,7 +14,7 @@ int main(){
         cin>>a[i];
     } 
 
-    int pre[n]; 
+    long long int pre[n]; 
     pre[0] = a[0]; 
 
     for(int i=1;i<n;i++){
